fix: refactor history page
