+++ src/components/historypage/recent-runs-card.tsx
@@ -38,10 +38,11 @@ export default function RecentRunsCard() {
 
   const recentRuns = useMemo(() => {
     return [...pastRuns]
-      .filter(run => 
-        run.achievedPace !== null && 
-        run.achievedDistance !== null && 
-        run.achievedDuration !== null
+      .filter(
+        (run) =>
+          run.achievedPace !== null &&
+          run.achievedDistance !== null &&
+          run.achievedDuration !== null,
       )
       .sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
       .slice(0, ITEMS_PER_PAGE);
